Change MaterializedRow calls to have explicit long
In preparation for INT/SMALLINT/TINYINT, it is required that the
MaterializedRow constructors do not assume all Numbers are to be
treated as longs. This changes all callsites of the row() method to
have explicit longs.

+++ presto-main/src/test/java/com/facebook/presto/operator/TestMarkDistinctOperator.java
@@ -79,7 +79,7 @@ public class TestMarkDistinctOperator
         Operator operator = operatorFactory.createOperator(driverContext);
 
         MaterializedResult.Builder expected = resultBuilder(driverContext.getSession(), BIGINT, BOOLEAN);
-        for (int i = 0; i < 100; i++) {
+        for (long i = 0; i < 100; i++) {
             expected.row(i, true);
             expected.row(i, false);
         }
